Tesseract and  Paddle separated, enhancement progress
